RAG_LangChain/requirements.txt
absl-py==2.1.0
accelerate==1.6.0
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.0
chronos-forecasting @ git+https://github.com/amazon-science/chronos-forecasting.git@6a9c8dadac04eb85befc935043e3e2cce914267f
click==8.1.8
cloudpickle==3.0.0
cmake==3.31.4
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
deepl==1.22.0
defusedxml==0.7.1
distro==1.9.0
dlib==19.24.6
docker==7.1.0
easyocr==1.7.2
emoji==2.14.1
eval_type_backport==0.2.0
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.115.12
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.16.1
Flask==3.1.0
fonttools==4.53.0
fqdn==1.5.1
fsspec==2024.10.0
googleapis-common-protos==1.70.0
grpcio==1.70.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
hf-xet==1.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.31.1
idna==3.10
imageio==2.37.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwipiepy==0.21.0
kiwipiepy_model==0.21.0
kiwisolver==1.4.5
langchain==0.3.25
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-teddynote==0.3.45
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
lazy_loader==0.4
lz4==4.4.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
mmh3==4.1.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
olefile==0.47
openai==1.82.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.3.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.10.0
platformdirs==4.2.2
prometheus_client==0.21.0
prompt==0.4.1
prompt_toolkit==3.0.47
protobuf==5.29.3
protoc-gen-openapiv2==0.0.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.0.3
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
safetensors==0.4.5
scikit-image==0.25.2
scipy==1.14.1
selenium==4.30.0
Send2Trash==1.8.3
setuptools==70.0.0
sgmllib3k==1.0.0
shapely==2.0.7
SimpleGUITk==1.1.3
simsimd==6.0.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.41
stable_baselines3==2.5.0
stack-data==0.6.3
starlette==0.46.2
stringzilla==3.10.8
sympy==1.13.1
tavily-python==0.7.2
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminado==0.18.1
tifffile==2025.3.30
tiktoken==0.9.0
timm==1.0.11
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.5.1
torchnet==0.0.4
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.51.3
trio==0.29.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241003
types-requests==2.32.0.20250515
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.34.2
visdom==0.2.4
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.43.0
widgetsnbextension==4.0.13
wsproto==1.2.0
xxhash==3.5.0
zstandard==0.23.0
